requirements.txt
Django
psycopg2
Pillow
pytest-django
faker